+++ src/main/java/com/example/ISTE240_Project/Models/WantToRead.java
@@ -10,7 +10,7 @@ public class WantToRead {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column
     private String isbn;
@@ -44,7 +44,7 @@ public class WantToRead {
     public WantToRead() {}
 
     // GETTERS
-    public Long getId() {return id;}
+    public int getId() {return id;}
     public String getIsbn() {return isbn;}
     public String getTitle() {return title;}
     public String getAuthor() {return author;}
@@ -57,7 +57,7 @@ public class WantToRead {
 
 
     // SETTERS
-    public void setId(Long id) {this.id = id;}
+    public void setId(int id) {this.id = id;}
     public void setIsbn(String isbn) {this.isbn = isbn;}
     public void setTitle(String title) {this.title = title;}
     public void setAuthor(String author) {this.author = author;}
